requirement.txt
streamlit
google-genai
pandas
python-dotenv